q-fin/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
akshare>=1.16
baostock>=0.8.8
tqdm>=4.65
anthropic>=0.40
openai>=1.50
duckduckgo-search>=6.0
python-dotenv>=1.0
